move the rhai engine management to its own module
the main file in the rhai module ios getting too large, because it
manages modules and functions registration, along with the plugin
interface. It should only take care of the plugin part
