+++ apollo-router/src/plugins/rhai/execution.rs
@@ -1,5 +1,5 @@
 //! execution module
 
 pub(crate) use crate::services::execution::*;
-pub(crate) type Response = super::RhaiExecutionResponse;
-pub(crate) type DeferredResponse = super::RhaiExecutionDeferredResponse;
+pub(crate) type Response = super::engine::RhaiExecutionResponse;
+pub(crate) type DeferredResponse = super::engine::RhaiExecutionDeferredResponse;
